feat: implement useEpisodes hook for fetching episode data
